clean up print component architecture
