+++ src/sidebar/components/tools/print/printRequest/wmtsJsonConfig/World_Imagery.json
@@ -1,0 +1,17 @@
+{
+    "type": "WMTS",
+    "baseURL": "https://server.arcgisonline.com/arcgis/rest/services/World_Imagery/MapServer/tile/{TileMatrix}/{TileRow}/{TileCol}",
+    "layer": "World_Imagery",
+    "imageFormat": "image/png",
+    "opacity": 1.0,
+    "style": "Default Style",
+    "version": "1.0.0",
+    "dimensions": [],
+    "dimensionParams": {},
+    "requestEncoding": "REST",
+    "customParams": {
+        "TRANSPARENT": "true"
+    },
+    "matrixSet": "EPSG:3857",
+    "matrices": []
+ }
