+++ src/sidebar/components/tools/print/printRequest/wmts_json_config_entries.js
