--- src/sidebar/components/tools/print/wmtsJsonConfig/Ortho_2016_Cache.json
@@ -1,101 +1,0 @@
-{
-    "type": "WMTS",
-    "baseURL": "https://maps.simcoe.ca/arcgis/rest/services/Public/Ortho_2016_Cache/MapServer/tile/{TileMatrix}/{TileRow}/{TileCol}",
-    "layer": "Ortho_2016_Cache",
-    "imageFormat": "image/png",
-    "opacity": 1.0,
-    "style": "Default Style",
-    "version": "1.0.0",
-    "dimensions": [],
-    "dimensionParams": {},
-    "requestEncoding": "REST",
-    "customParams": {
-        "TRANSPARENT": "true"
-    },
-    "matrixSet": "EPSG:3857",
-    "matrices": [{
-            "identifier": "0",
-            "scaleDenominator": 1091957.5469304253,
-            "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-            "tileSize": [256, 256],
-            "matrixSize": [144, 187]
-        },
-        {
-            "identifier": "1",
-            "scaleDenominator": 545978.7734656851,
-            "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-            "tileSize": [256, 256],
-            "matrixSize": [288, 373]
-        },
-        {
-            "identifier": "2",
-            "scaleDenominator": 272989.38673237007,
-            "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-            "tileSize": [256, 256],
-            "matrixSize": [575, 746]
-        },
-        {
-            "identifier": "3",
-            "scaleDenominator": 136494.69336618503,
-            "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-            "tileSize": [256, 256],
-            "matrixSize": [1149, 1491]
-        },
-        {
-            "identifier": "4",
-            "scaleDenominator": 68247.34668309252,
-            "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-            "tileSize": [256, 256],
-            "matrixSize": [2297, 2981]
-        },
-        {
-            "identifier": "5",
-            "scaleDenominator": 34123.67334154626,
-            "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-            "tileSize": [256, 256],
-            "matrixSize": [4594, 5962]
-        },
-        {
-            "identifier": "6",
-            "scaleDenominator": 17061.836671245605,
-            "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-            "tileSize": [256, 256],
-            "matrixSize": [9188, 11924]
-        },
-        {
-            "identifier": "7",
-            "scaleDenominator": 8530.918335622802,
-            "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-            "tileSize": [256, 256],
-            "matrixSize": [18376, 23848]
-        },
-        {
-            "identifier": "8",
-            "scaleDenominator": 4265.459167338929,
-            "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-            "tileSize": [256, 256],
-            "matrixSize": [36751, 47696]
-        },
-        {
-            "identifier": "9",
-            "scaleDenominator": 2132.729584141936,
-            "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-            "tileSize": [256, 256],
-            "matrixSize": [73502, 95391]
-        },
-        {
-            "identifier": "10",
-            "scaleDenominator": 1066.3647915984968,
-            "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-            "tileSize": [256, 256],
-            "matrixSize": [147003, 190781]
-        },
-        {
-            "identifier": "11",
-            "scaleDenominator": 532.9474944663505,
-            "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-            "tileSize": [256, 256],
-            "matrixSize": [294136, 381730]
-        }
-    ]
-}
