--- src/sidebar/components/tools/print/wmtsJsonConfig/World_Imagery.json
@@ -1,185 +1,0 @@
-{
-    "type": "WMTS",
-    "baseURL": "https://server.arcgisonline.com/arcgis/rest/services/World_Imagery/MapServer/tile/{TileMatrix}/{TileRow}/{TileCol}",
-    "layer": "World_Imagery",
-    "imageFormat": "image/png",
-    "opacity": 1.0,
-    "style": "Default Style",
-    "version": "1.0.0",
-    "dimensions": [],
-    "dimensionParams": {},
-    "requestEncoding": "REST",
-    "customParams": {
-        "TRANSPARENT": "true"
-    },
-    "matrixSet": "EPSG:3857",
-    "matrices": [{
-          "identifier": "0",
-          "scaleDenominator": 5.590822640285016E8,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [1, 1]
-       },
-       {
-          "identifier": "1",
-          "scaleDenominator": 2.7954113201425034E8,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [2, 2]
-       },
-       {
-          "identifier": "2",
-          "scaleDenominator": 1.3977056600712562E8,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [4, 4]
-       },
-       {
-          "identifier": "3",
-          "scaleDenominator": 6.988528300356235E7,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [8, 8]
-       },
-       {
-          "identifier": "4",
-          "scaleDenominator": 3.494264150178117E7,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [16, 16]
-       },
-       {
-          "identifier": "5",
-          "scaleDenominator": 1.7471320750890587E7,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [32, 32]
-       },
-       {
-          "identifier": "6",
-          "scaleDenominator": 8735660.375445293,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [64, 64]
-       },
-       {
-          "identifier": "7",
-          "scaleDenominator": 4367830.187722647,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [128, 128]
-       },
-       {
-          "identifier": "8",
-          "scaleDenominator": 2183915.0938617955,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [256, 256]
-       },
-       {
-          "identifier": "9",
-          "scaleDenominator": 1091957.5469304253,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [512, 512]
-       },
-       {
-          "identifier": "10",
-          "scaleDenominator": 545978.7734656851,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [1024, 1023]
-       },
-       {
-          "identifier": "11",
-          "scaleDenominator": 272989.38673237007,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [2048, 2045]
-       },
-       {
-          "identifier": "12",
-          "scaleDenominator": 136494.69336618503,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [4096, 4090]
-       },
-       {
-          "identifier": "13",
-          "scaleDenominator": 68247.34668309252,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [8192, 8179]
-       },
-       {
-          "identifier": "14",
-          "scaleDenominator": 34123.67334154626,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [16384, 16358]
-       },
-       {
-          "identifier": "15",
-          "scaleDenominator": 17061.836671245605,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [32768, 32715]
-       },
-       {
-          "identifier": "16",
-          "scaleDenominator": 8530.918335622802,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [65536, 65429]
-       },
-       {
-          "identifier": "17",
-          "scaleDenominator": 4265.459167338929,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [131072, 130858]
-       },
-       {
-          "identifier": "18",
-          "scaleDenominator": 2132.729584141936,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [262144, 261715]
-       },
-       {
-          "identifier": "19",
-          "scaleDenominator": 1066.3647915984968,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [524288, 523430]
-       },
-       {
-          "identifier": "20",
-          "scaleDenominator": 533.1823957992484,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [1048576, 1046859]
-       },
-       {
-          "identifier": "21",
-          "scaleDenominator": 266.5911978996242,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [2097152, 2093718]
-       },
-       {
-          "identifier": "22",
-          "scaleDenominator": 133.2955989498121,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [4194304, 4187435]
-       },
-       {
-          "identifier": "23",
-          "scaleDenominator": 66.64779947490605,
-          "topLeftCorner": [-2.0037508342787E7, 2.0037508342787E7],
-          "tileSize": [256, 256],
-          "matrixSize": [8388608, 8374869]
-       }
-    ]
- }
